Clean up, formatting and commenting
